Shard browser tests
This shards our browser tests to speed up CI runs. Test setup is
repeated for each shard but is small compared to actual test duration.

The monolithic `selenium.test.ts` is split into multiple files, since
jest will always assign an entire file to a shard.

This also adds a custom sequencer to make sure that two long tests do
not end up on the same shard.
