+++ src/frontend/src/test-e2e/delegationTtl.test.ts
@@ -1,0 +1,67 @@
+import { AuthenticateView, DemoAppView } from "./views";
+import { FLOWS } from "./flows";
+import {
+  addVirtualAuthenticator,
+  runInBrowser,
+  switchToPopup,
+  waitToClose,
+} from "./util";
+
+import { II_URL, TEST_APP_NICE_URL, DEVICE_NAME1 } from "./constants";
+
+test("Delegation maxTimeToLive: 1 min", async () => {
+  await runInBrowser(async (browser: WebdriverIO.Browser) => {
+    await addVirtualAuthenticator(browser);
+    const demoAppView = new DemoAppView(browser);
+    await demoAppView.open(TEST_APP_NICE_URL, II_URL);
+    await demoAppView.waitForDisplay();
+    expect(await demoAppView.getPrincipal()).toBe("2vxsx-fae");
+    await demoAppView.setMaxTimeToLive(BigInt(60_000_000_000));
+    await demoAppView.signin();
+    await switchToPopup(browser);
+    await FLOWS.registerNewIdentityAuthenticateView(DEVICE_NAME1, browser);
+    await waitToClose(browser);
+    await demoAppView.waitForDisplay();
+    expect(await demoAppView.getPrincipal()).not.toBe("2vxsx-fae");
+    const exp = await browser.$("#expiration").getText();
+    // compare only up to one decimal place for the 1min test
+    expect(Number(exp) / 60_000_000_000).toBeCloseTo(1, 0);
+  });
+}, 300_000);
+
+test("Delegation maxTimeToLive: 1 day", async () => {
+  await runInBrowser(async (browser: WebdriverIO.Browser) => {
+    await addVirtualAuthenticator(browser);
+    const demoAppView = new DemoAppView(browser);
+    await demoAppView.open(TEST_APP_NICE_URL, II_URL);
+    await demoAppView.waitForDisplay();
+    expect(await demoAppView.getPrincipal()).toBe("2vxsx-fae");
+    await demoAppView.setMaxTimeToLive(BigInt(86400_000_000_000));
+    await demoAppView.signin();
+    await switchToPopup(browser);
+    await FLOWS.registerNewIdentityAuthenticateView(DEVICE_NAME1, browser);
+    await waitToClose(browser);
+    expect(await demoAppView.getPrincipal()).not.toBe("2vxsx-fae");
+    const exp = await browser.$("#expiration").getText();
+    expect(Number(exp) / 86400_000_000_000).toBeCloseTo(1);
+  });
+}, 300_000);
+
+test("Delegation maxTimeToLive: 2 months", async () => {
+  await runInBrowser(async (browser: WebdriverIO.Browser) => {
+    await addVirtualAuthenticator(browser);
+    const demoAppView = new DemoAppView(browser);
+    await demoAppView.open(TEST_APP_NICE_URL, II_URL);
+    await demoAppView.waitForDisplay();
+    expect(await demoAppView.getPrincipal()).toBe("2vxsx-fae");
+    await demoAppView.setMaxTimeToLive(BigInt(5_184_000_000_000_000)); // 60 days
+    await demoAppView.signin();
+    await switchToPopup(browser);
+    await FLOWS.registerNewIdentityAuthenticateView(DEVICE_NAME1, browser);
+    await waitToClose(browser);
+    expect(await demoAppView.getPrincipal()).not.toBe("2vxsx-fae");
+    const exp = await browser.$("#expiration").getText();
+    // NB: Max out at 30 days
+    expect(Number(exp) / 2_592_000_000_000_000).toBeCloseTo(1);
+  });
+}, 300_000);
